src/py/DBCode/requirements.txt
annotated-types==0.7.0
anthropic==0.64.0
anyio==4.10.0
asyncclick==8.2.2.2
attrs==25.3.0
beautifulsoup4==4.13.5
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
contourpy==1.3.3
cycler==0.12.1
distro==1.9.0
docker==7.1.0
et_xmlfile==2.0.0
filelock==3.19.1
fonttools==4.59.2
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.40.3
google-genai==1.31.0
h11==0.16.0
hf-xet==1.1.8
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
idna==3.10
jiter==0.10.0
jsonpath-ng==1.7.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
kiwisolver==1.4.9
linkify-it-py==2.0.3
lxml==6.0.1
markdown-it-py==4.0.0
matplotlib==3.10.6
mcp==1.12.2
mdit-py-plugins==0.5.0
mdurl==0.1.2
numpy==2.3.2
ollama==0.5.3
openai==1.99.9
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
ply==3.11
ptyprocess==0.7.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rich==14.1.0
rpds-py==0.27.0
rsa==4.9.1
ruff==0.12.10
seaborn==0.13.2
setuptools==78.1.1
simhash==2.1.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.8
sqlglot==27.14.0
sse-starlette==3.0.2
starlette==0.47.2
tenacity==9.1.2
textual==5.3.0
tqdm==4.67.1
tree-sitter==0.21.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uc-micro-py==1.0.3
urllib3==2.5.0
uvicorn==0.35.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
wheel==0.45.1
